Feat: implement left apis for workbook detail page part of SWM-158 related to #61
