+++ src/app/(after-login)/(top)/workbookDetail/[workbookId]/workbook-detail-api.ts
@@ -1,4 +1,5 @@
 import { IWorkbookBase } from '@/model/i-workbook';
+import { IQuestionInWorkbook } from '@/model/i-question';
 
 export async function readWorkbookById(workbookId: string) {
   try {
@@ -52,5 +53,55 @@ export async function deleteWorkbook(workbookId: string) {
     return json.data;
   } catch (e) {
     console.error(`deleteWorkbook failed by ${e}`);
+  }
+}
+
+export async function deleteQuestionInWorkbook({
+  workbookId,
+  questionId,
+}: {
+  workbookId: string;
+  questionId: string;
+}) {
+  try {
+    const response = await fetch(
+      `/api/workbook/${workbookId}/question/${questionId}`,
+      {
+        method: 'DELETE',
+      },
+    );
+
+    return await response.json();
+  } catch (e) {
+    console.error(`deleteQuestionInWorkbook failed by ${e}`);
+  }
+}
+
+export async function updateQuestionSequence({
+  workbookId,
+  questions,
+}: {
+  workbookId: string;
+  questions: IQuestionInWorkbook[];
+}) {
+  const request = questions.map((ques, idx) => {
+    return {
+      questionSetId: ques.questionSetId,
+      sequence: `${idx + 1}`,
+    };
+  });
+
+  try {
+    const response = await fetch(
+      `/api/workbook/question-sequence/${workbookId}`,
+      {
+        method: 'PATCH',
+        body: JSON.stringify(request),
+      },
+    );
+
+    return await response.json();
+  } catch (e) {
+    console.error(`updateQuestionSequence failed by ${e}`);
   }
 }
